fix(iota-indexer): add serial_test for all test cases
refactor(iota-indexer): add common module for tests
