+++ crates/iota-indexer/tests/rpc-tests/main.rs
@@ -2,4 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #[cfg(feature = "pg_integration")]
+#[path = "../common/mod.rs"]
+mod common;
+
+#[cfg(feature = "pg_integration")]
 mod read_api;
